refactor(color-utility): creates files for components
